+++ relic-manager/src/ui/ui.py
@@ -1,4 +1,5 @@
 from ui.relics_view import RelicsView
+from ui.add_relic_view import AddRelicView
 
 
 class UI:
@@ -18,5 +19,11 @@ class UI:
     def _show_relics_view(self):
         self._hide_current_view()
 
-        self._current_view = RelicsView(self._root)
+        self._current_view = RelicsView(self._root, self._show_add_relic_view)
+        self._current_view.pack()
+
+    def _show_add_relic_view(self):
+        self._hide_current_view()
+
+        self._current_view = AddRelicView(self._root, self._show_relics_view)
         self._current_view.pack()